+++ core/lib/logger.js
@@ -1,19 +1,14 @@
 var debug = require('debug');
-
 module.exports = function(name) {
-  var log = debug(name);
-  var error = debug(name + ':error');
-
-  function logFn() {
-    log.apply(this, arguments);
-  }
-
-  function errorFn() {
-    error.apply(this, arguments);
-  }
+  var dlog = debug('theeye:log:' + name);
+  var derror = debug('theeye:error:' + name);
+  var ddata = debug('theeye:data:' + name);
+  var ddebug = debug('theeye:debug:' + name);
 
   return {
-    log: logFn,
-    error: errorFn
+    log   : function flog(){ dlog.apply(this, arguments); },
+    error : function ferror(){ derror.apply(this, arguments); },
+    data  : function fdata(){ ddata.apply(this, arguments); },
+    debug : function fdebug(){ ddebug.apply(this, arguments); }
   };
-};
+}