+++ core/main.js
@@ -1,11 +1,8 @@
 "use strict";
 
-var rootPath = require('app-root-path');
-require('app-root-path').setPath(rootPath + '/core');
-
 var ErrorHandler = require('./lib/errorHandler');
 
-var logger = require('./lib/logger')('eye:supervisor:main');
+var logger = require('./lib/logger')('main');
 logger.log('initializing supervisor');
 
 process.on('SIGINT', function(){